rename http module to client from service
